+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+-   Next occurrence calculation for scheduled transactions list was not considering modifications properly, leading to incorrect next occurrence dates.
+
 ## [0.14.2] - 2025-09-21
 
 ### Fixed